chore: moved sliding functions to the SlidingTab
